requirements.txt
# -----------------------------------------------------------------------------------
# Core dependencies
# -----------------------------------------------------------------------------------
discord.py~=2.6
python-dotenv~=1.0
aiohttp~=3.9
APScheduler~=3.10
requests~=2.32
pytz~=2024.1
psycopg2-binary~=2.9
tqdm~=4.66

# -----------------------------------------------------------------------------------
# Google Cloud
# -----------------------------------------------------------------------------------
google-cloud-firestore~=2.16
google-auth~=2.35
google-auth-oauthlib~=1.2
google-auth-httplib2~=0.2

# -----------------------------------------------------------------------------------
# AI / Gemini integration
# -----------------------------------------------------------------------------------
#google-generativeai~=0.6
#transformers~=4.45
#torch~=2.5
#sentencepiece~=0.2

# -----------------------------------------------------------------------------------
# Miscellaneous
# -----------------------------------------------------------------------------------
protobuf~=5.28
